Added tracking and simplified XML display.
Created nec-xml functions to simplify displaying XML content for NEC
devices, and rewrote associated existing functions.

Created tracking page that can be used as a banner message for IP
phones. This URI should be added to the XML home page in phone settings.
This will ensure detailed information on each phone, and will
later be used in a phone information API. This should not be used on a
DT7xx device, as they are old and it will not allow the home page to
close.

Moved device information to database, and added option to populate
variables into database from command line with --databaseInit